Added Ball class, Added Collision class
In this update rendering images was fixed, Previously rendering images
gaves errors where the whole texture wouldnt be rendered to the screen.
Ball class was added with temporary code to make it bounce from one side
of the screen too another, Collision class was started and is the next
thing to get finished.