Added a Main menu
Now the main menu is working and the game is about 90% done. I want to
add sound and some effects when the ball hits the paddle. I wanted to
add a more advanced main menu but it just game me more problems than
satisfaction. It would only work on my development computer and not any
other device it was tested on.